requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
pandas==2.2.2
requests==2.32.3
rapidfuzz==3.9.7
python-dotenv==1.0.1
apache-airflow==2.9.3
gunicorn==22.0.0
